+++ next.config.js
@@ -1,8 +1,3 @@
-const path = require('path')
-
 module.exports = {
-  sassOptions: {
-    includePaths: [path.join(__dirname, 'styles')],
-  },
   reactStrictMode: true,
 }
