+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "total-seg-next",
+  "name": "total-seg-next-build",
   "version": "0.1.0",
   "private": true,
   "scripts": {
@@ -16,15 +16,12 @@
     "nodemailer": "^6.6.3",
     "react": "17.0.2",
     "react-dom": "17.0.2",
-    "react-hook-form": "^7.15.3",
-    "sass": "^1.41.0",
-    "swiper": "^7.0.6"
+    "react-hook-form": "^7.15.3"
   },
   "devDependencies": {
-    "babel-eslint": "^10.1.0",
-    "eslint": "^7.2.0",
-    "eslint-config-airbnb": "^18.2.1",
+    "eslint": "7.32.0",
     "eslint-config-next": "11.1.2",
+    "eslint-config-airbnb": "^18.2.1",
     "eslint-config-node": "^4.1.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-import-resolver-webpack": "^0.13.1",
@@ -34,6 +31,7 @@
     "eslint-plugin-prettier": "^4.0.0",
     "eslint-plugin-react": "^7.21.5",
     "eslint-plugin-react-hooks": "^1.7.0",
+    "babel-eslint": "^10.1.0",
     "prettier": "^2.4.1"
   }
 }
