+++ pages/_document.js
@@ -8,7 +8,7 @@ class MyDocument extends Document {
 
   render() {
     return (
-      <Html lang="pt-BR">
+      <Html lang="pt-BR" style={{ scrollBehavior: "smooth" }}>
         <Head />
         <body>
           <Main />
