+++ pages/area-do-cliente.jsx
@@ -1,0 +1,3 @@
+export default function ClientArea() {
+  return <div>Area do cliente</div>;
+}
