--- pages/index.js
@@ -1,612 +1,0 @@
-// next components
-import Head from "next/head";
-import Image from "next/image";
-import Link from "next/link";
-// dependencies
-import { useForm } from "react-hook-form";
-import axios from "axios";
-// components
-import Slide from "../components/slider/Slider";
-import CTAButton from "../components/CTAButton";
-import SiteMenu from "../components/Menu";
-import Card from "../components/Card";
-// theme
-import theme from "../styles/Theme";
-// data
-import websiteData from "../data/data";
-
-export default function Home() {
-  const {
-    register,
-    handleSubmit,
-    watch,
-    formState: { errors },
-  } = useForm();
-
-  const weCallSubmit = (data) => {
-    axios.post("/api", data).then(console.log);
-  };
-
-  return (
-    <>
-      <div>
-        <Head>
-          <title>Create Next App</title>
-          <meta name="description" content="Generated by create next app" />
-          <link rel="icon" href="/favicon.ico" />
-        </Head>
-
-        {/* ================= Nav =============================== */}
-
-        <SiteMenu />
-        <main>
-          {/* Header Section */}
-          <header className="container" id="#inicio">
-            <div className="header__main__info">
-              <h3 className="header__subtitle">O melhor</h3>
-              <h1 className="header__title">
-                Alarme monitorado
-                <br />
-                <span className="header__title__span">
-                  {" "}
-                  Com aviso a polícia
-                </span>
-              </h1>
-            </div>
-
-            <div className="header__subinfo">
-              <h2 className="hader__subinfo__heading">
-                <span>100%</span> Sem fio, com
-              </h2>
-              <h2 className="hader__subinfo__heading">
-                <span>50%</span> de desconto
-              </h2>
-              <small className="small__text">por tempo limitado</small>
-            </div>
-
-            <CTAButton />
-
-            <p className="header__description">
-              A união perfeita entre a ultima geração de câmeras de segurança
-              com gravação em full HD e um sistema de alarme monitorado de alta
-              precisão <span>com aviso a policia.</span>
-            </p>
-          </header>
-
-          {/* Section call-us */}
-          <section className="we__call">
-            <div className="we__call__Info container">
-              <h2>Prefere que entremos em contato?</h2>
-              <p>
-                Informe um telefone que um de nosssos especialistas entrará em
-                contato
-              </p>
-
-              <form onSubmit={handleSubmit(weCallSubmit)}>
-                <input
-                  type="text"
-                  placeholder="Digite seu nome"
-                  {...register("nome")}
-                />
-                <input
-                  type="phone"
-                  placeholder="Digite seu telefone"
-                  {...register("telefone")}
-                />
-                <button className="send__button" type="submit">
-                  Enviar
-                </button>
-              </form>
-            </div>
-            <div className="we__call__Image"></div>
-          </section>
-
-          {/* section advantages */}
-          <section className="advantages" id="#vantagens">
-            <div className="advantages__text__wrapper">
-              <h2>Principais Vantagens</h2>
-              <p>
-                Com o alarme monitorado total segurança, seu patrimonio está
-                sempre protegido{" "}
-              </p>
-            </div>
-
-            <div className="advantages__swiper">
-              <Slide>
-                {websiteData.advantagesCarroussel.map((item, index) => (
-                  <div className="keen-slider__slide slider__item">
-                    <div className="slide__image">
-                      <Image
-                        src={item.image}
-                        width={313}
-                        height={192}
-                        alt={item.alt}
-                      />
-                    </div>
-                    <div className="slider__title">Total Segurança</div>
-                    <div className="slider__subtitle">{item.subtitle}</div>
-                    <div className="slider__description">
-                      {item.description}
-                    </div>
-                  </div>
-                ))}
-              </Slide>
-              <div className="button__wrapper">
-                <CTAButton />
-              </div>
-            </div>
-
-            <div className="advanges__card">
-              {websiteData.advantagesCards.map((item) => (
-                <Card data={item} />
-              ))}
-            </div>
-          </section>
-
-          {/* Section App */}
-          <section className="app__container container">
-            <h2>Controle tudo através do seu smartphone</h2>
-            <p>
-              Através de nosso aplicativo você gerencia o seu alarme e acompanha
-              acontece em sua empresa ou residência em tempo real com áudio
-              bilateral e imagem de alta qualidade. Tudo isso de qualquer lugar
-              no mundo através do seu smartphone.
-            </p>
-            <div className="app__image__container">
-              <Image
-                src={"/mobile-app-image.webp"}
-                width={400}
-                height={315}
-                alt={"Aplicativo Total Segurança"}
-              />
-            </div>
-            <div className="button__wrapper">
-              <CTAButton />
-            </div>
-          </section>
-
-          {/* Section Testimonials */}
-          <section className="testimonials">
-            <div className="testimonial__text__wrapper">
-              <h2>+ de 7.500 clientes</h2>
-              <p>
-                {" "}
-                Confiam a segurança de seus comércios e residencias ao alarme
-                monitrado total segurança
-              </p>
-            </div>
-
-            <div className="testimonial__swiper">
-              <Slide>
-                {websiteData.testimonials.map((item, index) => (
-                  <div className="keen-slider__slide slider__item">
-                    <div className="testimonial__wrapper">
-                      <div className="testimonial__image">
-                        <Image
-                          src={item.image}
-                          width={44}
-                          height={44}
-                          alt={item.alt}
-                        />
-                      </div>
-                      <h2 className="testimonial__title">{item.client}</h2>
-                      <h3 className="testimonial__subtitle">{item.whois}</h3>
-                      <p className="testimonial__description">
-                        {item.testimonial}
-                      </p>
-                    </div>
-                    <Link href="/area-do-cliente">
-                      <a className="testimonial__add__link">
-                        Envie seu depoimento
-                      </a>
-                    </Link>
-                  </div>
-                ))}
-              </Slide>
-            </div>
-          </section>
-
-          {/* Section Equipments */}
-          <section className="equipment container">
-            <div className="equipment__text__wrapper">
-              <h2>Conheça seu novo equipamento</h2>
-              <p>teste de equipamento</p>
-            </div>
-
-            {websiteData.equipment.map((item, index) => (
-              <div className="equipment__content" key={index}>
-                <Image
-                  src={item.image}
-                  alt={item.alt}
-                  width={375}
-                  height={346}
-                />
-
-                <div className="equipment__content-text">
-                  <h2>{item.title}</h2>
-                  <p>{item.description}</p>
-
-                  {index % 2 === 0 ? (
-                    <div className="button__wrapper">
-                      <CTAButton />
-                    </div>
-                  ) : (
-                    ""
-                  )}
-                </div>
-              </div>
-            ))}
-          </section>
-
-          {/* Section call-us */}
-          <section className="we__call">
-            <div className="we__call__Info container">
-              <h2>Prefere que entremos em contato?</h2>
-              <p>
-                Informe um telefone que um de nosssos especialistas entrará em
-                contato
-              </p>
-
-              <form onSubmit={handleSubmit(weCallSubmit)}>
-                <input
-                  type="text"
-                  placeholder="Digite seu nome"
-                  {...register("nome")}
-                />
-                <input
-                  type="phone"
-                  placeholder="Digite seu telefone"
-                  {...register("telefone")}
-                />
-                <button className="send__button" type="submit">
-                  Enviar
-                </button>
-              </form>
-            </div>
-            <div className="we__call__Image"></div>
-          </section>
-        </main>
-
-        <footer className="container">
-          <Image
-            src="/web-logo-total-seguranca.svg"
-            alt="Total segurança Logo"
-            width={134}
-            height={43}
-          />
-
-          <div className="footer__info">Total segurança © 2021</div>
-          <div className="footer__links">
-            <Link href="/termos-de-serviço">
-              <a className="footer__link">Termos de serviço</a>
-            </Link>
-
-            <Link href="/termos-de-privacidade">
-              <a className="footer__link">Termos de privacidade</a>
-            </Link>
-          </div>
-        </footer>
-      </div>
-
-      {/* Global document style */}
-      <style jsx>{`
-        .container {
-          padding: 4rem 1.2rem;
-          overflow: hidden;
-        }
-      `}</style>
-
-      {/* Header style */}
-      <style jsx>{`
-        header {
-          margin-top: 3rem;
-        }
-
-        .header__main__info {
-          display: grid;
-          gap: 0.8rem;
-        }
-
-        .header__title {
-          font-size: 3rem;
-          font-weight: 800;
-          color: ${theme.color.mainColor};
-          line-height: 90%;
-        }
-
-        .header__title span {
-          color: #0b0b0b;
-          font-weight: 600;
-          font-size: 1.9rem;
-        }
-
-        .header__subtitle {
-          font-size: 1.1rem;
-          font-weight: 700;
-        }
-
-        .header__subinfo {
-          margin-top: 1rem;
-          display: grid;
-          gap: 0.4rem;
-        }
-
-        .hader__subinfo__heading {
-          font-size: 1.6rem;
-          font-weight: 800;
-        }
-
-        .hader__subinfo__heading span {
-          color: ${theme.color.mainColor};
-        }
-
-        .small__text {
-          font-size: 0.8rem;
-        }
-
-        .header__description {
-          margin-top: 2rem;
-          color: #0b0b0b;
-          font-size: 1rem;
-          line-height: 130%;
-        }
-        .header__description span {
-          color: ${theme.color.mainColor};
-        }
-      `}</style>
-
-      {/* We call section */}
-      <style jsx>{`
-
-        .we__call {
-          transform: skewY(-3deg);
-          z-index: 0;
-
-        }
-
-        .we__call__Info {
-          border-radius: .5rem;
-          background: ${theme.color.mainLinear}};
-        }
-
-        .we__call__Info h2{
-          transform: skewY(3deg);
-          font-size: 2.5rem;
-          margin-bottom: 1rem;
-          font-weight: 800;
-          color: white;
-        }
-
-        .we__call__Info p{
-          transform: skewY(3deg);
-          display: block;
-          margin-bottom: 1.5rem;
-          color: white;
-        }
-
-        .we__call__Info form{
-          transform: skewY(3deg);
-          display: grid;
-          gap: 1rem;
-          max-width: 20rem;
-        }
-
-        .we__call__Info input{
-          height: 3.125rem;
-          border-radius: 2rem;
-          padding-left: 1.5rem;
-          font-size: .95rem;
-          background-color: rgba(255,255,255,.75);
-        }
-
-        .we__call__Info input::placeholder{
-          color: #000;
-        }
-
-         .send__button{
-          height: 3.125rem;
-          border-radius: 2rem;
-          font-size: .95rem;
-
-          background: none;
-          outline: none;
-          border: none;
-
-          background-color: white;
-         
-        }
-
-        @media(min-width: 1280px){
-          .send__button{
-            border: 1px solid white;
-          color: white;
-          }
-        }
-
-      `}</style>
-
-      {/* section advantages section */}
-      <style jsx>{`
-        .advantages {
-          padding: 4rem 0;
-        }
-
-        .advantages__text__wrapper {
-          padding: 0 1.2rem 2rem;
-        }
-
-        .slider__item {
-          padding: 0rem 1.5rem;
-        }
-
-        .slider__title {
-          color: #ff0015;
-          font-weight: 800;
-          font-size: 20px;
-        }
-
-        .slide__image {
-          margin-bottom: 1.8rem;
-        }
-
-        .slider__subtitle {
-          font-weight: 800;
-          font-size: 23px;
-          color: #0b0b0b;
-          margin-bottom: 1.5rem;
-        }
-
-        .slider__description {
-          font-size: 1rem;
-          line-height: 130%;
-          color: ${theme.color.mainTextColor};
-        }
-
-        .button__wrapper {
-          max-width: 17rem;
-          margin: 0 auto;
-        }
-
-        .advanges__card {
-          display: grid;
-          grid-template-columns: auto auto;
-          gap: 1rem;
-          margin-top: 2.5rem;
-          padding: 0 1.2rem;
-        }
-      `}</style>
-
-      {/* Section App */}
-      <style jsx>{`
-        .app__container {
-        }
-
-        .app__container h2 {
-          font-size: 2.5rem;
-          line-height: 110%;
-          font-weight: 800;
-          color: ${theme.color.mainColor};
-          margin-bottom: 1.5rem;
-        }
-
-        .app__container p {
-          font-size: 1rem;
-          line-height: 130%;
-          color: ${theme.color.mainTextColor};
-        }
-
-        .app__image__container {
-          display: grid;
-          padding: 1.5rem 0 0;
-        }
-      `}</style>
-
-      {/* Section Testimonials] */}
-      <style jsx>{`
-        .testimonial {
-          padding: 4rem 0 0;
-        }
-
-        .testimonial__text__wrapper h2 {
-          font-size: 1.8rem;
-          font-weight: 700;
-          color: #ff0015;
-          margin-bottom: 1rem;
-        }
-
-        .testimonial__swiper {
-          padding: 1rem 0;
-        }
-
-        .testimonial__wrapper {
-          width: 300px;
-          background: #ff0015;
-          box-shadow: 0px 15px 30px rgba(0, 0, 0, 0.25);
-          border-radius: 1rem;
-          padding: 1.5rem;
-          display: grid;
-          gap: 0.4rem;
-          margin-bottom: 1rem;
-        }
-
-        .testimonial__title {
-          font-size: 1, 0625rem.;
-          color: #fff;
-        }
-
-        .testimonial__subtitle {
-          font-size: 0, 8125rem;
-          color: #ffa3a3;
-        }
-
-        .testimonial__description {
-          font-size: 0, 9rem;
-          color: #ffffff80;
-          line-height: 130%;
-        }
-
-        .testimonial__text__wrapper {
-          padding: 1.2rem;
-        }
-
-        .testimonial__add__link {
-          font-size: 0.813;
-          font-weight: 600;
-          color: ${theme.color.mainColor};
-          padding: 0 1rem;
-
-          margin-bottom: 2rem;
-          display: block;
-        }
-      `}</style>
-
-      {/* Section equipment */}
-      <style jsx>{`
-        .equipment__text__wrapper {
-        }
-
-        .equipment__content {
-          padding: 3rem 0;
-        }
-
-        .equipment__content-text {
-          margin-top: 2rem;
-          text-align: center;
-          padding: 0 1rem;
-        }
-        .equipment__content-text h2 {
-          text-align: center;
-          font-weight: 800;
-          font-size: 23px;
-          color: ${theme.color.mainColor};
-          margin-bottom: 1.5rem;
-        }
-
-        .equipment__content-text p {
-          color: #0b0b0b;
-          font-size: 1rem;
-        }
-      `}</style>
-
-      {/* Footer */}
-      <style jsx>{`
-        footer {
-          display: grid;
-          gap: 2rem;
-          justify-content: center;
-          align-items: center;
-        }
-
-        .footer__info {
-          text-align: center;
-        }
-
-        .footer__link {
-          font-size: 0.813rem;
-          color: #00000099;
-          padding: 1rem;
-        }
-      `}</style>
-    </>
-  );
-}
