+++ pages/total-cotacao.jsx
@@ -1,0 +1,3 @@
+export default function ClientArea() {
+  return <div>Cotação</div>;
+}
